+++ NEWS.md
@@ -5,7 +5,8 @@
 ##### (Release date: XXXXXXX)
 
 ## Changes and additions:
-* selections longer than 20 s can be analyzed with specan function
+*Spectrogram creating functions (manualoc, specreator, trackfreqs, dfts, ffts, dfDTW, ffDTW, lspec) can performe faster (4-10 times) using a different image algorithm. This can be called with the argument "fast". Note that not all spectrogram options are available when fast = TRUE 
+* selections longer than 20 s can be analyzed with specan function (previously it returned an error, this is going to be slow anyways)
 * "clip.edges" argument in compare.methods function to set removing mising values at the start and end of signals when using dfDTW and ffDTW methods  
 * "treshold" argument in compare.methods function to set the amplitude detection threshold for ffDTW, dfDTW and SP methods  
 * "exclude" agument in imp.syrinx function to exclude selection files that cannot be read